refactor(source-panel): extract reusable mountSourcePanel from concept-create.js
Moves the Text/URL/File tab panel markup and all handlers (tab switching,
paste textarea, URL input, file dropzone, Attach/Cancel buttons) from
concept-create.js::beginEditSource into public/js/source-panel.js.
beginEditSource now delegates to mountSourcePanel(valueEl, {onAttach, onCancel})
and the two callbacks preserve byte-equivalent behavior for state.source writes,
emitTelemetry, and rerenderSummary calls. No HTML structure or class names changed.
